=str #16028 Cleaned up ducplicate implementations of FlowOps
